+++ src/services/project.service.js
@@ -5,6 +5,45 @@ export default class ProjectService extends Service{
         super('projects', axios)
     }
 
+    getCurrentUserAuthorities(id, then, error) {
+        return this.axios({
+            url: this.controllerName + '/' + id + '/authorities',
+            method: 'get',
+          })
+          .then(resp => {
+            then(resp.data)
+          })
+          .catch(err => {
+              error(err);
+          });
+    }
+
+    getCurrentUserMailSubscription(projectId, then, error) {
+        return this.axios({
+            url: this.controllerName + '/' + projectId + '/my-mail-subscription',
+            method: 'get',
+        })
+        .then(resp => {
+            then(resp.data)
+        })
+        .catch(err => {
+            error(err);
+        });
+    }
+
+    setCurrentUserMailSubscription(projectId, mailEnabled, then, error) {
+        return this.axios({
+            url: this.controllerName + '/' + projectId + '/my-mail-subscription/' + mailEnabled,
+            method: 'put',
+        })
+        .then(resp => {
+            then(resp.data)
+        })
+        .catch(err => {
+            error(err);
+        });
+    }
+
     create(name, domain, contractId, then, error){
         const project = {
             'name' : name,